+++ src/seleksipegawai/laman_pegawai.java
@@ -53,21 +53,16 @@ public class laman_pegawai extends javax.swing.JFrame {
         for (int i = 0; i < baris; i++) {
             defaultTableModel.removeRow(i);
         }
-
-
         String sql = "select id_pegawai,nama, jk, tgl_lahir, alamat from pegawai";
         try {
             statement = (Statement) con.createStatement();
             resultSet = statement.executeQuery(sql);
-
             while (resultSet.next()) {
                 String kolom1 = resultSet.getString(1);
                 String kolom2 = resultSet.getString(2);
                 String kolom3 = resultSet.getString(3);
                 String kolom4 = resultSet.getString(4);
                 String kolom5 = resultSet.getString(5);
-
-
                 String kolom[] = {kolom1, kolom2,kolom3,kolom4,kolom5};
                 defaultTableModel.addRow(kolom);
             }
@@ -77,16 +72,17 @@ public class laman_pegawai extends javax.swing.JFrame {
         }
     
   }
- 
-  
+
 //    /**
 //     * Creates new form laman_pegawai
 //     */
     public laman_pegawai() {
+        koneksi1 koneksi = new koneksi1();
+        con = (Connection) koneksi.koneksi();
         initComponents();
-         auto_number();
-         update_tabel();
-        kosongkan_text(); 
+        auto_number();
+        update_tabel();
+        kosongkan_text();
        
     }
     
@@ -130,53 +126,39 @@ public class laman_pegawai extends javax.swing.JFrame {
 //        }
 //    }
     
-     private void kosongkan_text(){
-           id_peg.setText("");
+    private void kosongkan_text() {
+        id_peg.setText("");
         nm_peg.setText("");
         jk.setSelectedItem(null);
         tglhr.setDate(null);
         alamat_peg.setText("");
         auto_number();
-       }
+    }
      
- private void koneksi() {
+ 
+    private void auto_number() {
         try {
-            Class.forName("com.mysql.jdbc.Driver");
-            con = (Connection) DriverManager.getConnection("jdbc:mysql://localhost/pegawai1", "root", "");
-
+            String sql = "SELECT MAX(right(id_pegawai,2))AS no from pegawai";
+            statement = (Statement) con.createStatement();
+            ResultSet rs = statement.executeQuery(sql);
+            while (rs.next()) {
+                if (rs.first() == false) {
+                    id_peg.setText("P001");
+                } else {
+                    rs.last();
+                    int auto_id = rs.getInt(1) + 1;
+                    String no = String.valueOf(auto_id);
+                    int noLong = no.length();
+                    for (int a = 0; a < 3 - noLong; a++) {
+                        no = "0" + no;
+                    }
+                    id_peg.setText("P" + no);
+                }
+            }
         } catch (Exception e) {
+            JOptionPane.showMessageDialog(this, "Error:\n" + e.toString(), "Kesalahan", JOptionPane.WARNING_MESSAGE);
         }
     }
-   private void auto_number(){
-       
- try {
-   koneksi();
-           String sql ="SELECT MAX(right(id_pegawai,2))AS no from pegawai";
-             statement = (Statement) con.createStatement();
-          ResultSet rs= statement.executeQuery(sql);
-         while (rs.next()) {
-               if (rs.first() == false){
-                id_peg.setText("P001");
-               }else{
-                    rs.last();
-                   int auto_id = rs.getInt(1) + 1;
-                  String no = String.valueOf(auto_id);
-                   int noLong=no.length();
-                   for (int a = 0; a< 3-noLong; a++){
-                       no= "0" + no;
-                   }
-                   id_peg.setText("P" + no);
-                   
-               }
-           }
-     } catch (Exception e) {
-           JOptionPane.showMessageDialog(this, "Error:\n" + e.toString(), "Kesalahan", JOptionPane.WARNING_MESSAGE);
-
-     }
-       }
-           
-        
-  
     /**
      * This method is called from within the constructor to initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is always
@@ -211,7 +193,7 @@ public class laman_pegawai extends javax.swing.JFrame {
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
 
-        jInternalFrame1.setTitle("Pegawai");
+        jInternalFrame1.setTitle("Master Pegawai");
         jInternalFrame1.setVisible(true);
 
         jToolBar1.setRollover(true);
@@ -420,7 +402,7 @@ public class laman_pegawai extends javax.swing.JFrame {
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 424, Short.MAX_VALUE)
+                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 436, Short.MAX_VALUE)
                 .addGap(29, 29, 29))
         );
 
@@ -441,74 +423,63 @@ public class laman_pegawai extends javax.swing.JFrame {
     }// </editor-fold>//GEN-END:initComponents
 
     private void jTable1MouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jTable1MouseClicked
-
-     try {
-  int row =jTable1.getSelectedRow();
-       String tabel_klik;
-          tabel_klik = (jTable1.getModel().getValueAt(row, 0).toString());
-       java.sql.Connection conn =(java.sql.Connection)seleksipegawai.koneksi1.koneksiDB();
-           java.sql.Statement stm = conn.createStatement();
-        java.sql.ResultSet sql = stm.executeQuery("select * from pegawai where id_pegawai='"+tabel_klik+"'");
-          if(sql.next()){
-              String add1= sql.getString("id_pegawai");
-              id_peg.setText(add1);
-             String add2 = sql.getString("nama");
-               nm_peg.setText(add2);
+        try {
+            int row = jTable1.getSelectedRow();
+            String tabel_klik;
+            tabel_klik = (jTable1.getModel().getValueAt(row, 0).toString());
+            java.sql.Connection conn = (java.sql.Connection) seleksipegawai.koneksi1.koneksiDB();
+            java.sql.Statement stm = conn.createStatement();
+            java.sql.ResultSet sql = stm.executeQuery("select * from pegawai where id_pegawai='" + tabel_klik + "'");
+            if (sql.next()) {
+                String add1 = sql.getString("id_pegawai");
+                id_peg.setText(add1);
+                String add2 = sql.getString("nama");
+                nm_peg.setText(add2);
                 String add3 = sql.getString("jk");
-               jk.getSelectedItem();
+                jk.getSelectedItem();
                 Date add4 = sql.getDate("tgl_lahir");
-               tglhr.setDate(add4);
-              String add5 = sql.getString("alamat");
-          alamat_peg.setText(add5);
-             
-         }     }
-            catch (Exception e) {
+                tglhr.setDate(add4);
+                String add5 = sql.getString("alamat");
+                alamat_peg.setText(add5);
             }
-//        
-        
-//        id_peg.setText(jTable1.getValueAt(jTable1.getSelectedRow(), 1).toString());
-//     nm_peg.setText(jTable1.getValueAt(jTable1.getSelectedRow(), 2).toString());
-//        jk.(jTable1.getValueAt(jTable1.getSelectedRow(), 3).toString());
+        } catch (Exception e) {
+        }
     }//GEN-LAST:event_jTable1MouseClicked
 
     private void jButton2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton2ActionPerformed
-       java.util.Date tanggal= (java.util.Date) this.tglhr.getDate();
-         String jeniska="L";
+        java.util.Date tanggal = (java.util.Date) this.tglhr.getDate();
+        String jeniska = "L";
         if (jk.getSelectedItem().equals("Laki-laki")) {
-            jeniska="L";
+            jeniska = "L";
         } else {
-            jeniska= "P";
+            jeniska = "P";
         }
-        
-         String nama="",alamat="";
-         if(nm_peg.getText().equals(""))
-     {
-         JOptionPane.showConfirmDialog(null, "Nama Pegawai Tidak Boleh Kosong");
-        
-     }else if(alamat_peg.getText().equals("")) 
-     {
-         JOptionPane.showConfirmDialog(null,"Alamat Pegawai Tidak Boleh Kosong" );
-     }else
-     {
-         try {
-            String sql;
-            sql = "insert into pegawai values('"+id_peg.getText()+"','"+nm_peg.getText()+"','"+jeniska+"','"+new java.sql.Date(tanggal.getTime())+"','"+alamat_peg.getText()+"')";
-            System.out.println(sql);
-            java.sql.Connection conn = (java.sql.Connection) seleksipegawai.koneksi1.koneksiDB();
-            java.sql.PreparedStatement pst = conn.prepareStatement(sql);
-            pst.execute();
-            JOptionPane.showMessageDialog(null, "Berhasil Disimpan");
+        String nama = "", alamat = "";
+        if (nm_peg.getText().equals("")) {
+            JOptionPane.showConfirmDialog(null, "Nama Pegawai Tidak Boleh Kosong");
+        } else if (alamat_peg.getText().equals("")) {
+            JOptionPane.showConfirmDialog(null, "Alamat Pegawai Tidak Boleh Kosong");
+        } else {
+            try {
+                int dialogButton = JOptionPane.YES_NO_OPTION;
+                int dialogResult = JOptionPane.showConfirmDialog(this, "Your Message", "Title on Box", dialogButton);
+                if (dialogResult == 0) {
+                    String sql;
+                    sql = "insert into pegawai values('" + id_peg.getText() + "','" + nm_peg.getText() + "','" + jeniska + "','" + new java.sql.Date(tanggal.getTime()) + "','" + alamat_peg.getText() + "')";
+                    System.out.println(sql);
+                    java.sql.Connection conn = (java.sql.Connection) seleksipegawai.koneksi1.koneksiDB();
+                    java.sql.PreparedStatement pst = conn.prepareStatement(sql);
+                    pst.execute();
+                    JOptionPane.showMessageDialog(null, "Berhasil Disimpan");
+                } else {
+                }
+            } catch (Exception e) {
+                JOptionPane.showMessageDialog(null, e);
+            }
+            update_tabel();
+            auto_number();
+            kosongkan_text();
         }
-        catch (Exception e) {
-            JOptionPane.showMessageDialog(null, e);
-        }
-        
-        update_tabel();
-        auto_number();
-        kosongkan_text();
-//        tampilanadministrator();
-       
-     }   
     }//GEN-LAST:event_jButton2ActionPerformed
 
     private void jButton3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton3ActionPerformed
@@ -516,28 +487,26 @@ public class laman_pegawai extends javax.swing.JFrame {
         if (jk.getSelectedItem().equals("Laki-laki")) {
             jeniska="L";
         } else {
-            jeniska= "P";
+            jeniska="P";
         }
-//        try {
+        try {
             String value1 = id_peg.getText();
             String value2 = nm_peg.getText();
             java.util.Date tanggal = (java.util.Date) this.tglhr.getDate();
             String value4=alamat_peg.getText();
-            String sql;
-            sql ="update pegawai set  nama='"+value2+"',jk='"+jeniska+"', tgl_lahir='"+new java.sql.Date(tanggal.getTime())+"',alamat='"+value4+"' where id_pegawai='"+value1+"'";
+            String sql ="update pegawai set  nama='"+value2+"',jk='"+jeniska+"', tgl_lahir='"+new java.sql.Date(tanggal.getTime())+"',alamat='"+value4+"' where id_pegawai='"+value1+"'";
             System.out.println(sql);
-//            java.sql.Connection conn;
-//            conn = (java.sql.Connection) seleksipegawai.koneksi1.koneksiDB();
-//            java.sql.PreparedStatement pst = conn.prepareStatement(sql);
-//            pst.execute();
-//            JOptionPane.showMessageDialog(null, "Edit ?");
-//        } catch (Exception e) {          
-//    JOptionPane.showMessageDialog(null, "Error");
-//        }  
-//        update_tabel();
-//        auto_number();
-//        kosongkan_text();
-
+            java.sql.Connection conn;
+            conn = (java.sql.Connection) seleksipegawai.koneksi1.koneksiDB();
+            java.sql.PreparedStatement pst = conn.prepareStatement(sql);
+            pst.execute();
+            JOptionPane.showMessageDialog(null, "Edit ?");
+        } catch (Exception e) {          
+    JOptionPane.showMessageDialog(null, "Error");
+        }  
+        update_tabel();
+        auto_number();
+        kosongkan_text();
     }//GEN-LAST:event_jButton3ActionPerformed
 
     private void jButton5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton5ActionPerformed
@@ -554,31 +523,32 @@ public class laman_pegawai extends javax.swing.JFrame {
     }//GEN-LAST:event_jButton5ActionPerformed
 
     private void jButton4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton4ActionPerformed
-  try {
-            String sql ="delete from pegawai where id_pegawai=? ";
+        try {
+            int dialogButton = JOptionPane.YES_NO_OPTION;
+            int dialogResult = JOptionPane.showConfirmDialog(this, "Hapus data ini?", "Title on Box", dialogButton);
+            if (dialogResult == 0) {
+            String sql = "delete from pegawai where id_pegawai=? ";
             java.sql.Connection conn = (java.sql.Connection) seleksipegawai.koneksi1.koneksiDB();
             java.sql.PreparedStatement pst = conn.prepareStatement(sql);
-
             pst.setString(1, id_peg.getText());
             pst.execute();
-
-            JOptionPane.showMessageDialog(null, "Hapus");
-           
-
-        } catch (Exception e) {
-        }
+            JOptionPane.showMessageDialog(null, "Berhasil Dihapus");
+             } else {
+                }
+            } catch (Exception e) {
+                JOptionPane.showMessageDialog(null, e);
+            }
         update_tabel();
         auto_number();
         kosongkan_text();
-        // TODO add your handling code here:
     }//GEN-LAST:event_jButton4ActionPerformed
 
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
-  dispose();
+        dispose();
     }//GEN-LAST:event_jButton1ActionPerformed
 
     private void jButton7ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton7ActionPerformed
-kosongkan_text();        
+        kosongkan_text();        
     }//GEN-LAST:event_jButton7ActionPerformed
 
     private void id_pegPropertyChange(java.beans.PropertyChangeEvent evt) {//GEN-FIRST:event_id_pegPropertyChange
